# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smenka API",
    "description": "API для учёта рабочего времени (shift tracking).\n\n## Режимы работы\n\n- **Персональный** — любой авторизованный пользователь может трекать рабочее время для себя.\n- **Организационный** — владелец создаёт организацию, приглашает сотрудников, настраивает правила (геопроверка, лимиты пауз, автозавершение).\n\n## Роли в организации\n\n| Роль | Описание | Может трекать время | Управление |\n|------|----------|---------------------|------------|\n| **Owner** | Создатель организации | Нет | Полное: настройки, участники, статистика, смены сотрудников |\n| **Admin** | Участник с расширенными правами | Да | Рабочие точки, просмотр смен и статистики |\n| **Employee** | Обычный участник | Да | Только свои смены |\n\n> **Важно:** Owner НЕ является участником организации и не может трекать в ней время. Это управленческая роль.\n\n## Формат ответов\n\nВсе ответы обёрнуты в единую структуру:\n```json\n{\"data\": <payload>, \"error\": null}\n```\nПри ошибке:\n```json\n{\"data\": null, \"error\": {\"code\": \"ERROR_CODE\", \"message\": \"Описание ошибки\"}}\n```\n\n## Авторизация\n\nИспользуется JWT Bearer-токен. Получите `access_token` через `/auth/login` и передавайте в заголовке:\n```\nAuthorization: Bearer <access_token>\n```\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/organizations/{org_id}/members/{user_id}/checklists": {
      "get": {
        "tags": [
          "checklist-assignments"
        ],
        "summary": "Эффективные чек-листы сотрудника",
        "description": "Вычисляет итоговый набор чек-листов по формуле: (шаблоны роли − remove) + add. Доступно владельцу, админам и самому сотруднику.",
        "operationId": "get_member_effective_checklists_api_v1_organizations__org_id__members__user_id__checklists_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/shifts/{shift_id}/checklists": {
      "get": {
        "tags": [
          "checklist-instances"
        ],
        "summary": "Чек-листы смены",
        "description": "Список экземпляров чек-листов смены со сводкой. Доступно владельцу смены, владельцу и админам организации.",
        "operationId": "list_shift_checklists_api_v1_shifts__shift_id__checklists_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Shift Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ChecklistInstanceListResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shifts/{shift_id}/checklists/{instance_id}": {
      "get": {
        "tags": [
          "checklist-instances"
        ],
        "summary": "Детали экземпляра чек-листа",
        "description": "Экземпляр с упорядоченными пунктами.",
        "operationId": "get_instance_api_v1_shifts__shift_id__checklists__instance_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Shift Id"
            }
          },
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Instance Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ChecklistInstanceDetailResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shifts/{shift_id}/checklists/{instance_id}/items/{item_id}": {
      "patch": {
        "tags": [
          "checklist-instances"
        ],
        "summary": "Обновить пункт",
        "description": "Отметить/снять пункт и добавить комментарий. Только владелец смены, только пока смена активна.",
        "operationId": "update_item_api_v1_shifts__shift_id__checklists__instance_id__items__item_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Shift Id"
            }
          },
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Instance Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Код ошибки (e.g. SHIFT_NOT_FOUND, FORBIDDEN)"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Человекочитаемое описание ошибки"
          },
          "validation": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation",
            "description": "Детали ошибок валидации полей"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError"
      },
      "ApiResponse": {
        "properties": {
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Полезная нагрузка ответа (null при ошибке)"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ],
            "description": "Описание ошибки (null при успехе)"
          }
        },
        "type": "object",
        "title": "ApiResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InstanceItemUpdate": {
        "properties": {
          "is_completed": {
            "type": "boolean",
            "title": "Is Completed",
            "description": "Отмечено/снято"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "is_completed"
        ],
        "title": "InstanceItemUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "ApiResponse_ChecklistInstanceDetailResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChecklistInstanceDetailResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Полезная нагрузка ответа (null при ошибке)"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ],
            "description": "Описание ошибки (null при успехе)"
          }
        },
        "type": "object",
        "title": "ApiResponse[ChecklistInstanceDetailResponse]"
      },
      "ApiResponse_ChecklistInstanceListResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChecklistInstanceListResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Полезная нагрузка ответа (null при ошибке)"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ],
            "description": "Описание ошибки (null при успехе)"
          }
        },
        "type": "object",
        "title": "ApiResponse[ChecklistInstanceListResponse]"
      },
...
      "ChecklistInstanceDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "is_required": {
            "type": "boolean",
            "title": "Is Required"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "max_photos_per_item": {
            "type": "integer",
            "title": "Max Photos Per Item",
            "description": "Лимит фото на пункт (= CHECKLIST_MAX_PHOTOS_PER_ITEM)"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InstanceItemResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "organization_timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Timezone",
            "description": "Текущая IANA-таймзона организации смены; null для персональной смены"
          },
          "fill_allowed": {
            "type": "boolean",
            "title": "Fill Allowed",
            "description": "Можно ли сейчас редактировать чек-лист смены (checklist_grace_period): true для активной смены и для завершённой в открытом окне дозаполнения, false после его истечения",
            "default": true
          },
          "fill_deadline_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fill Deadline At",
            "description": "Момент закрытия окна дозаполнения в UTC; null для активной смены или уже закрытого окна (checklist_grace_period)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "is_required",
          "status",
          "completed_at",
          "created_at",
          "max_photos_per_item",
          "items"
        ],
        "title": "ChecklistInstanceDetailResponse"
      },
      "ChecklistInstanceListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChecklistInstanceResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "organization_timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Timezone",
            "description": "Текущая IANA-таймзона организации смены; null для персональной смены"
          },
          "fill_allowed": {
            "type": "boolean",
            "title": "Fill Allowed",
            "description": "Можно ли сейчас редактировать чек-листы смены (checklist_grace_period): true для активной смены и для завершённой в открытом окне дозаполнения, false после его истечения",
            "default": true
          },
          "fill_deadline_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fill Deadline At",
            "description": "Момент закрытия окна дозаполнения в UTC; null для активной смены или уже закрытого окна (checklist_grace_period)"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ChecklistInstanceListResponse"
      },
      "ChecklistInstanceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "is_required": {
            "type": "boolean",
            "title": "Is Required"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "pending | completed | incomplete"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "items_summary": {
            "$ref": "#/components/schemas/ItemsSummary"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "is_required",
          "status",
          "completed_at",
          "items_summary",
          "created_at"
        ],
        "title": "ChecklistInstanceResponse"
      },
      "InstanceItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "is_required": {
            "type": "boolean",
            "title": "Is Required"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "is_completed": {
            "type": "boolean",
            "title": "Is Completed"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "change_count": {
            "type": "integer",
            "title": "Change Count"
          },
          "photo_requirement": {
            "type": "string",
            "title": "Photo Requirement"
          },
          "photo_source": {
            "type": "string",
            "title": "Photo Source"
          },
          "photos_count": {
            "type": "integer",
            "title": "Photos Count",
            "default": 0
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/PhotoResponse"
            },
            "type": "array",
            "title": "Photos"
          }
        },
        "type": "object",
        "required": [
          "id",
          "text",
          "is_required",
          "position",
          "is_completed",
          "comment",
          "completed_at",
          "change_count",
          "photo_requirement",
          "photo_source"
        ],
        "title": "InstanceItemResponse"
      },
      "ItemsSummary": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "completed": {
            "type": "integer",
            "title": "Completed",
            "description": "Пунктов с is_completed=true (без учёта фото)"
          },
          "satisfied_count": {
            "type": "integer",
            "title": "Satisfied Count",
            "description": "Пунктов, прошедших критерий satisfied (с учётом обязательных фото)"
          },
          "photos_required_missing": {
            "type": "integer",
            "title": "Photos Required Missing",
            "description": "Пунктов с photo_requirement=required и без фото (бейдж «нужно фото»)"
          }
        },
        "type": "object",
        "required": [
          "total",
          "completed",
          "satisfied_count",
          "photos_required_missing"
        ],
        "title": "ItemsSummary"
      },
...
      "PhotoResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "file_id": {
            "type": "string",
            "title": "File Id"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url",
            "description": "Свежий presigned GET URL; null при недоступности storage"
          },
          "url_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url Expires At"
          },
          "captured_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "id",
          "file_id",
          "position"
        ],
        "title": "PhotoResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}